+++ src/features/Conversation/components/MarkdownElements/LobeThinking/Render.tsx
@@ -4,6 +4,8 @@ import Thinking from '@/components/Thinking';
 import { ARTIFACT_THINKING_TAG } from '@/const/plugin';
 import { useChatStore } from '@/store/chat';
 import { chatSelectors } from '@/store/chat/selectors';
+import { useUserStore } from '@/store/user';
+import { userGeneralSettingsSelectors } from '@/store/user/selectors';
 
 import { MarkdownElementProps } from '../type';
 import { isTagClosed } from '../utils';
@@ -13,12 +15,14 @@ const Render = memo<MarkdownElementProps>(({ children, id }) => {
     const message = chatSelectors.getMessageById(id)(s);
     return [!isTagClosed(ARTIFACT_THINKING_TAG, message?.content)];
   });
+  const transitionMode = useUserStore(userGeneralSettingsSelectors.transitionMode);
 
   return (
     <Thinking
       content={children as string}
       style={{ width: isGenerating ? '100%' : undefined }}
       thinking={isGenerating}
+      thinkingAnimated={transitionMode === 'fadeIn' && isGenerating}
     />
   );
 });